+++ .github/workflows/cross-repo-adev-docs.yml
@@ -1,4 +1,4 @@
-name: Update ADEV Angular CDK APIs and CLI Help
+name: Update ADEV Cross Repo Docs
 
 on:
   workflow_dispatch:
@@ -13,8 +13,8 @@ permissions:
   contents: read
 
 jobs:
-  update_adev_docs:
-    name: Update Angular CDK/Aria APIs and CLI Help (if necessary)
+  update_adev_cross_repo_docs:
+    name: Update Cross Repo ADEV Docs
     if: github.repository == 'angular/angular'
     runs-on: ubuntu-latest
     # Prevents multiple concurrent runs of this workflow for the same branch
@@ -31,89 +31,29 @@ jobs:
           # an authenticated URL.
           persist-credentials: false
 
-      # Angular CDK
-      - name: Generate CDK apis
-        run: node adev/scripts/update-cdk-apis/index.mjs
+      - name: Generate Docs
+        run: node adev/scripts/update-cross-repo-docs/index.mjs
         env:
-          ANGULAR_CDK_BUILDS_READONLY_GITHUB_TOKEN: ${{ secrets.READONLY_GITHUB_TOKEN }}
-      - name: Create a PR CDK apis (if necessary)
+          ANGULAR_READONLY_GITHUB_TOKEN: ${{ secrets.READONLY_GITHUB_TOKEN }}
+
+      - name: Create a PR (if necessary)
         uses: peter-evans/create-pull-request@271a8d0340265f705b14b6d32b9829c1cb33d45e # v7.0.8
         with:
           token: ${{ secrets.ANGULAR_ROBOT_ACCESS_TOKEN }}
           push-to-fork: 'angular-robot/angular'
           delete-branch: true
           maintainer-can-modify: false
-          branch: docs-update-cdk-apis-${{github.ref_name}}
+          branch: docs-cross-repo-update-${{github.ref_name}}
           committer: Angular Robot <angular-robot@google.com>
           author: Angular Robot <angular-robot@google.com>
-          title: 'docs: update Angular CDK apis [${{github.ref_name}}]'
+          title: 'docs: update cross-repo adev docs [${{github.ref_name}}]'
           body: |
-            Updated Angular CDK api files.
+            Updated Angular adev cross repo docs files.
           labels: |
             action: merge
             area: docs
             target: automation
           commit-message: |
-            docs: update Angular CDK apis
+            docs: update cross-repo adev docs
 
-            Updated Angular CDK api files.
-
-      # Angular CLI
-      - name: Reset to current GitHub ref
-        run: |
-          git reset --hard ${{ github.ref }}
-      - name: Generate CLI help
-        run: node adev/scripts/update-cli-help/index.mjs
-        env:
-          ANGULAR_CLI_BUILDS_READONLY_GITHUB_TOKEN: ${{ secrets.READONLY_GITHUB_TOKEN }}
-      - name: Create a PR CLI help (if necessary)
-        uses: peter-evans/create-pull-request@271a8d0340265f705b14b6d32b9829c1cb33d45e # v7.0.8
-        with:
-          token: ${{ secrets.ANGULAR_ROBOT_ACCESS_TOKEN }}
-          push-to-fork: 'angular-robot/angular'
-          delete-branch: true
-          maintainer-can-modify: false
-          branch: docs-update-cli-help-${{github.ref_name}}
-          committer: Angular Robot <angular-robot@google.com>
-          author: Angular Robot <angular-robot@google.com>
-          title: 'docs: update Angular CLI help [${{github.ref_name}}]'
-          body: |
-            Updated Angular CLI help contents.
-          labels: |
-            action: merge
-            area: docs
-            target: automation
-          commit-message: |
-            docs: update Angular CLI help
-
-            Updated Angular CLI help contents.
-
-      # Angular Aria
-      - name: Reset to current GitHub ref
-        run: |
-          git reset --hard ${{ github.ref }}
-      - name: Generate Aria apis
-        run: node adev/scripts/update-aria-apis/index.mjs
-        env:
-          ANGULAR_ARIA_BUILDS_READONLY_GITHUB_TOKEN: ${{ secrets.READONLY_GITHUB_TOKEN }}
-      - name: Create a PR Aria apis (if necessary)
-        uses: peter-evans/create-pull-request@271a8d0340265f705b14b6d32b9829c1cb33d45e # v7.0.8
-        with:
-          token: ${{ secrets.ANGULAR_ROBOT_ACCESS_TOKEN }}
-          push-to-fork: 'angular-robot/angular'
-          delete-branch: true
-          maintainer-can-modify: false
-          branch: docs-update-aria-apis-${{github.ref_name}}
-          committer: Angular Robot <angular-robot@google.com>
-          author: Angular Robot <angular-robot@google.com>
-          title: 'docs: update Angular Aria apis [${{github.ref_name}}]'
-          body: |
-            Updated Angular Aria api files.
-          labels: |
-            action: merge
-            area: docs
-            target: automation
-          commit-message: |
-            docs: update Angular Aria apis
-
-            Updated Angular Aria api files.
+            Updated Angular adev cross repo docs files.
