--- adev/scripts/update-aria-apis/README.md
@@ -1,3 +1,0 @@
-# Generating data for `angular.dev/api` Aria packages
-
-This script updates the Angular Aria api JSON files stored in `adev/src/content/aria`. This files are used to generate the [angular.dev api](https://angular.dev/api) pages for the Aria packages.
