--- adev/scripts/update-cdk-apis/README.md
@@ -1,3 +1,0 @@
-# Generating data for `angular.dev/api` CDK packages
-
-This script updates the Angular CDK api JSON files stored in `adev/src/content/cdk`. This files are used to generate the [angular.dev api](https://angular.dev/api) pages for the CDK packages.
