--- adev/scripts/update-cli-help/README.md
@@ -1,3 +1,0 @@
-# Generating data for `angular.dev/cli`
-
-This script updates the Angular CLI help JSON files stored in `adev/src/content/cli/help`. This files are used to generate the [angular.dev CLI](https://angular.dev/cli) pages.
