+++ adev/scripts/update-cross-repo-docs/README.md
@@ -1,0 +1,15 @@
+# Update cross-repository documentation assets
+
+This script updates documentation assets from other repositories for `angular.dev`.
+
+## Aria Packages
+
+Updates the Angular Aria API JSON files in `adev/src/content/aria`, used for the [angular.dev API](httpss.angular.dev/api) pages.
+
+## CDK Packages
+
+Updates the Angular CDK API JSON files in `adev/src/content/cdk`, used for the [angular.dev API](httpss.angular.dev/api) pages.
+
+## CLI Help Pages
+
+Updates the Angular CLI help JSON files in `adev/src/content/cli/help`, used for the [angular.dev CLI](httpss.angular.dev/cli) pages.
